+++ CHANGELOG.md
@@ -10,4 +10,8 @@
 - Documentation updates
 ## [1.0.3] - 2024-07-29
 - Added the SceneCamera class for easy camera access
+## [1.0.4] - 2024-08-01
+- Reworked how Gizmos are drawn in the CameraState class
+- Removed the local flag from the CameraState class
+- Added RectTransform support to the TrackerCameraModule